chore: remove unnecessary class
